@@ -18,21 +18,21 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.2
-  provider: ^4.3.1
-  pigment: ^1.0.3
-  http_interceptor: ^0.3.2
-  webview_flutter: ^2.0.2 
+  http: ^0.13.4
+  provider: ^6.0.1
+  pigment: ^1.0.4
+  http_interceptor: ^1.0.2
+  webview_flutter: ^2.1.2 
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
