@@ -2,25 +2,25 @@ import 'dart:async' show Future;
 import 'package:http/http.dart' as http;
 
 class ServiceBase {
-  final String apiBaseUrl;
+  final String? apiBaseUrl;
 
-  static ServiceBase _singleton;
+  static ServiceBase? _singleton;
 
   factory ServiceBase({apiBaseUrl}) {
     _singleton ??= ServiceBase._(apiBaseUrl: apiBaseUrl);
-    return _singleton;
+    return _singleton!;
   }
 
   ServiceBase._({this.apiBaseUrl});
 
-  static String getApiBaseUrl() {
-    return _singleton.apiBaseUrl;
+  static String? getApiBaseUrl() {
+    return _singleton!.apiBaseUrl;
   }
 
   static Future<http.Response> get({
-    String url,
+    required String url,
   }) async {
-    final response = await http.get(url);
+    final response = await http.get(Uri.parse(url));
     return response;
   }
 }
