@@ -4,7 +4,7 @@ import 'package:webview_flutter/webview_flutter.dart';
 class WebViewScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    var data = ModalRoute.of(context).settings.arguments as Map;
+    var data = ModalRoute.of(context)!.settings.arguments as Map;
     return WebView(
       javascriptMode: JavascriptMode.unrestricted,
       initialUrl: data["link"],
