@@ -13,12 +13,12 @@ import 'package:flutter/material.dart';
 class Navigation {
   Navigation._();
   static dynamic pushNamed(BuildContext context, String routeName,
-      [Object arguments]) {
+      [Object? arguments]) {
     Navigator.of(context).pushNamed(routeName, arguments: arguments);
   }
 
   static dynamic pushReplacementNamed(BuildContext context, String routeName,
-      [Object arguments]) {
+      [Object? arguments]) {
     Navigator.of(context).pushReplacementNamed(
       routeName,
       arguments: arguments,
@@ -26,7 +26,7 @@ class Navigation {
   }
 
   static dynamic pushAndRemoveUntil(BuildContext context, String routeName,
-      [Object arguments]) {
+      [Object? arguments]) {
     Navigator.of(context).pushNamedAndRemoveUntil(
         routeName, (Route<dynamic> route) => false,
         arguments: arguments);
@@ -38,7 +38,7 @@ class Navigation {
 
   /// [popAndPushNamed] - Pop the current route off the navigator and push a named route in its place.
   static dynamic popAndPushNamed(BuildContext context, String routeName,
-      [Object arguments]) {
+      [Object? arguments]) {
     Navigator.of(context).popAndPushNamed(routeName, arguments: arguments);
   }
 
